requirements.txt
peakdetect==1.1
Django==3.2.8
djangorestframework==3.12.4
django-cors-headers==3.10.0
tensorflow-cpu==2.6.0
openpyxl==3.0.9
pandas==1.3.3
scikit-learn==1.0
PyJWT==2.2.0
python-environ==0.4.54
mysqlclient==2.0.3
gunicorn==20.1.0